Update Validate-NDESConfiguration.ps1
Get-WmiObject is depracated 
(https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.management/get-wmiobject?view=powershell-5.1)

This caused an issue when running with PowerShell 7 as a I received:
The term 'Get-WmiObject' is not recognized as a name of a cmdlet, function, script file, or executable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again. 

Replaced:
(Get-WmiObject Win32_ComputerSystem).domain

With:
(Get-ADDomain).DNSRoot